Updated enet. Updated Gaff. Fixed a lot of VS2013 warnings in OtterUI. Added ModelLoader and ModelComponent. Changed the GetAllocator() API to return a pointer instead of a reference. To avoid potential null dereferencing. Removed m4 source generators. I never use them. Added RasterManager.

+++ dependencies/OtterUI/src/Graphics/Graphics.cpp
@@ -263,7 +263,7 @@ namespace Otter
 			DrawBatch& lastBatch = mBatches[mBatches.size() - 1];
 			if(	primType == Otter::kPrim_TriangleList && 
 				lastBatch.mPrimitiveType == primType && 
-				lastBatch.mTextureID == textureID &&
+				(uint32)lastBatch.mTextureID == textureID &&
 				lastBatch.mRenderFlags == renderFlags &&
 				!(!mPreTransformVerts && mStackUpdated) &&
 				!mPropertiesUpdated)
